// src/openapi/model/updatePatrolSelfOfAutolivCommand.ts
/**
 * ICNC.ERP.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { IDItemsPair } from './iDItemsPair';
import { PatrolSelfType } from './patrolSelfType';


export interface UpdatePatrolSelfOfAutolivCommand { 
    product_id?: string;
    machine_code?: string | null;
    id_items_pairs?: Array<IDItemsPair> | null;
    patrol_self_type?: PatrolSelfType;
}
export namespace UpdatePatrolSelfOfAutolivCommand {
}